♻️ 코드 정리.
- 위치 변경.
- 이름 변경.
- 미사용 코드 삭제.
