+++ app/src/debug/java/com/github/hemoptysisheart/parking/app/ui/preview/PreviewModel.kt
@@ -1,9 +1,9 @@
 package com.github.hemoptysisheart.parking.app.ui.preview
 
 import com.github.hemoptysisheart.parking.core.client.google.dto.TransportationMode
-import com.github.hemoptysisheart.parking.core.model.GeoSearchModel
-import com.github.hemoptysisheart.parking.core.model.GeoSearchModelImpl
 import com.github.hemoptysisheart.parking.core.model.LocationModel
+import com.github.hemoptysisheart.parking.core.model.LocationModelImpl
+import com.github.hemoptysisheart.parking.core.model.SensorModel
 import com.github.hemoptysisheart.parking.core.model.dto.PlaceSearchResult
 import com.github.hemoptysisheart.parking.core.model.dto.RouteSearchResult
 import com.github.hemoptysisheart.parking.domain.GeoLocation
@@ -11,15 +11,15 @@ import com.github.hemoptysisheart.parking.domain.Location
 import com.github.hemoptysisheart.parking.domain.Overview
 
 object PreviewModel {
-    val LOCATION_MODEL = object : LocationModel {
+    val LOCATION_MODEL = object : SensorModel {
         override val location: GeoLocation = GeoLocation(37.5638354, 126.9040477)
 
-        override fun addCallback(key: Any, callback: (GeoLocation) -> Unit) {}
+        override fun addLocationCallback(key: Any, callback: (GeoLocation) -> Unit) {}
 
-        override fun removeCallback(key: Any) {}
+        override fun removeLocationCallback(key: Any) {}
     }
 
-    val GEO_SEARCH_MODEL = object : GeoSearchModel {
+    val GEO_SEARCH_MODEL = object : LocationModel {
         override suspend fun searchDestination(center: GeoLocation, query: String): PlaceSearchResult {
             return PlaceSearchResult(center, query, listOf(), null)
         }
@@ -38,7 +38,7 @@ object PreviewModel {
             return RouteSearchResult(
                 origin,
                 destination,
-                GeoSearchModelImpl.TRANSPORTATION_MODE_MAP[mode]!!,
+                LocationModelImpl.TRANSPORTATION_MODE_MAP[mode]!!,
                 Overview(listOf())
             )
         }
