--- app/src/main/java/com/github/hemoptysisheart/parking/app/ui/configuration/Constant.kt
@@ -1,17 +1,0 @@
-package com.github.hemoptysisheart.parking.app.ui.configuration
-
-@Deprecated(
-    "use TAG_COMPOSE",
-    replaceWith = ReplaceWith("com.github.hemoptysisheart.parking.app.configuration.ConstantsKt.TAG_COMPOSE")
-)
-object Constant {
-    /**
-     * `@Composable` 함수용 태그.
-     */
-    const val TAG_COMPOSE = "Compose"
-
-    /**
-     * `com.github.hemoptysisheart.parking.app.ui.support` 패키지용.
-     */
-    const val TAG_SUPPORT = "Support"
-}
