+++ app/src/main/java/com/github/hemoptysisheart/parking/app/ui/molecule/search/RecommendedItem.kt
@@ -1,7 +1,5 @@
-package com.github.hemoptysisheart.parking.app.ui.component
+package com.github.hemoptysisheart.parking.app.ui.molecule.search
 
-import android.annotation.SuppressLint
-import android.util.Log
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.Text
@@ -12,21 +10,20 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.hemoptysisheart.parking.app.domain.distance
 import com.github.hemoptysisheart.parking.app.ui.atom.Distance
-import com.github.hemoptysisheart.parking.app.ui.configuration.Constant.TAG_COMPOSE
 import com.github.hemoptysisheart.parking.app.ui.preview.RecommendItems.ITEM_株式会社ＡＡＡ
 import com.github.hemoptysisheart.parking.domain.GeoLocation
 import com.github.hemoptysisheart.parking.domain.Location
 import com.github.hemoptysisheart.parking.domain.RecommendItem
 import com.github.hemoptysisheart.parking.ui.theme.ParkingTheme
+import com.github.hemoptysisheart.parking.app.ui.support.LOGGER_COMPOSE as LOGGER
 
 @Composable
-fun MapRecommendedItem(
+fun RecommendedItem(
     item: RecommendItem<*>,
     here: Location = GeoLocation(0.0, 0.0),
-    distanceCalculator: (Location) -> Double = { 12345.0 },
     onSelect: (RecommendItem<*>) -> Unit = {}
 ) {
-    Log.v(TAG_COMPOSE, "#MapRecommendedItem args : item=$item, onSelectRecommend=$onSelect")
+    LOGGER.v("#MapRecommendedItem args : item=$item, onSelectRecommend=$onSelect")
 
     Column(
         modifier = Modifier
@@ -51,9 +48,8 @@ fun MapRecommendedItem(
 
 @Composable
 @Preview(showBackground = true)
-@SuppressLint("ComposableNaming")
-fun preview_MapRecommendedItem_GmpPlace() {
+fun Preview_MapRecommendedItem_GmpPlace() {
     ParkingTheme {
-        MapRecommendedItem(ITEM_株式会社ＡＡＡ)
+        RecommendedItem(ITEM_株式会社ＡＡＡ)
     }
 }
