+++ app/src/main/java/com/github/hemoptysisheart/parking/app/ui/page/MainPage.kt
@@ -15,7 +15,7 @@ import com.github.hemoptysisheart.parking.app.ui.template.main.HeaderTemplate
 import com.github.hemoptysisheart.parking.app.ui.template.main.MapTemplate
 import com.github.hemoptysisheart.parking.app.viewmodel.MainViewModel
 import com.github.hemoptysisheart.parking.ui.theme.ParkingTheme
-import com.github.hemoptysisheart.parking.app.configuration.LOGGER_COMPOSE as LOGGER
+import com.github.hemoptysisheart.parking.app.ui.support.LOGGER_COMPOSE as LOGGER
 
 @Composable
 fun MainPage(
