+++ app/src/main/java/com/github/hemoptysisheart/parking/app/ui/page/SearchPage.kt
@@ -15,7 +15,7 @@ import com.github.hemoptysisheart.parking.app.ui.template.search.SearchHeaderTem
 import com.github.hemoptysisheart.parking.app.ui.template.search.SearchResultTemplate
 import com.github.hemoptysisheart.parking.app.viewmodel.SearchViewModel
 import com.github.hemoptysisheart.parking.ui.theme.ParkingTheme
-import com.github.hemoptysisheart.parking.app.configuration.LOGGER_COMPOSE as LOGGER
+import com.github.hemoptysisheart.parking.app.ui.support.LOGGER_COMPOSE as LOGGER
 
 @Composable
 fun SearchPage(
