--- app/src/main/java/com/github/hemoptysisheart/parking/app/ui/state/MainScreenState.kt
@@ -1,10 +1,0 @@
-package com.github.hemoptysisheart.parking.app.ui.state
-
-import com.github.hemoptysisheart.parking.app.ui.state.OverlayState.COLLAPSE
-
-data class MainScreenState(
-    /**
-     * 추천 항목을 포함하는 오버레이를 열어서 보일지 여부.
-     */
-    var overlayState: OverlayState = COLLAPSE
-)
