--- app/src/main/java/com/github/hemoptysisheart/parking/app/ui/state/OverlayState.kt
@@ -1,18 +1,0 @@
-package com.github.hemoptysisheart.parking.app.ui.state
-
-enum class OverlayState {
-    /**
-     * 숨기기.
-     */
-    HIDE,
-
-    /**
-     * 접기(헤더만 표시)
-     */
-    COLLAPSE,
-
-    /**
-     * 열기(전체화면)
-     */
-    EXTEND;
-}
