+++ app/src/main/java/com/github/hemoptysisheart/parking/app/ui/support/constants.kt
@@ -1,6 +1,9 @@
-package com.github.hemoptysisheart.parking.app.configuration
+package com.github.hemoptysisheart.parking.app.ui.support
 
 import com.github.hemoptysisheart.parking.core.util.Logger
 
 const val TAG_COMPOSE = "Compose"
 val LOGGER_COMPOSE = Logger(TAG_COMPOSE)
+
+const val TAG_SUPPORT = "Support"
+val LOGGER_SUPPORT = Logger(TAG_SUPPORT)
