+++ app/src/main/java/com/github/hemoptysisheart/parking/app/ui/template/main/HeaderTemplate.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.zIndex
 import com.github.hemoptysisheart.parking.ui.theme.ParkingTheme
-import com.github.hemoptysisheart.parking.app.configuration.LOGGER_COMPOSE as LOGGER
+import com.github.hemoptysisheart.parking.app.ui.support.LOGGER_COMPOSE as LOGGER
 
 @Composable
 fun HeaderTemplate(
