+++ app/src/main/java/com/github/hemoptysisheart/parking/app/viewmodel/SearchViewModel.kt
@@ -2,8 +2,8 @@ package com.github.hemoptysisheart.parking.app.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.github.hemoptysisheart.parking.core.model.GeoSearchModel
 import com.github.hemoptysisheart.parking.core.model.LocationModel
+import com.github.hemoptysisheart.parking.core.model.SensorModel
 import com.github.hemoptysisheart.parking.core.util.Logger
 import com.github.hemoptysisheart.parking.domain.Location
 import com.github.hemoptysisheart.parking.domain.RecommendItem
@@ -15,19 +15,19 @@ import javax.inject.Inject
 
 @HiltViewModel
 class SearchViewModel @Inject constructor(
-    private val locationModel: LocationModel,
-    private val geoSearchModel: GeoSearchModel
+    private val sensorModel: SensorModel,
+    private val locationModel: LocationModel
 ) : ViewModel() {
     companion object {
         private const val TAG = "SearchViewModel"
         private val LOGGER = Logger(TAG)
     }
 
-    var here: Location = locationModel.location
-        get() = locationModel.location
-        private set
-
     private var searchJob: Job? = null
+
+    var here: Location = sensorModel.location
+        get() = sensorModel.location
+        private set
 
     val query = MutableStateFlow("")
 
@@ -43,9 +43,12 @@ class SearchViewModel @Inject constructor(
         searchJob?.cancel()
         searchJob = viewModelScope.launch {
             resultList.emit(
-                geoSearchModel.searchDestination(locationModel.location, it)
+                locationModel.searchDestination(sensorModel.location, it)
                     .places
             )
         }
     }
+
+    override fun toString() = "$TAG(sensorModel=$sensorModel, locationModel=$locationModel, " +
+            "query=${query.value}, resultList=${resultList.value})"
 }
