--- core/src/main/java/com/github/hemoptysisheart/parking/core/model/GeoSearchModelImpl.kt
@@ -1,133 +1,0 @@
-package com.github.hemoptysisheart.parking.core.model
-
-import android.util.Log
-import com.github.hemoptysisheart.parking.core.client.google.DirectionsParams
-import com.github.hemoptysisheart.parking.core.client.google.MapsClient
-import com.github.hemoptysisheart.parking.core.client.google.NearbySearchParams
-import com.github.hemoptysisheart.parking.core.client.google.dto.PlaceDescriptor
-import com.github.hemoptysisheart.parking.core.client.google.dto.PlaceTypes
-import com.github.hemoptysisheart.parking.core.client.google.dto.RankBy
-import com.github.hemoptysisheart.parking.core.client.google.dto.TransportationMode
-import com.github.hemoptysisheart.parking.core.logging.logArgs
-import com.github.hemoptysisheart.parking.core.model.dto.LocationGmpPlace
-import com.github.hemoptysisheart.parking.core.model.dto.PlaceSearchResult
-import com.github.hemoptysisheart.parking.core.model.dto.RouteSearchResult
-import com.github.hemoptysisheart.parking.domain.*
-import com.github.hemoptysisheart.util.TimeProvider
-import java.time.Instant
-
-class GeoSearchModelImpl(
-    private val mapsClient: MapsClient,
-    private val timeProvider: TimeProvider
-) : GeoSearchModel {
-    companion object {
-        private const val TAG = "GeoSearchModelImpl"
-
-        /**
-         * 목적지 주변 주차장 검색 반경 기본값. meter 단위.
-         */
-        const val SEARCH_PARKING_RADIUS = 100
-
-        val TRANSPORTATION_MODE_MAP = mapOf(
-            TransportationMode.WALKING to Transport.WALK,
-            TransportationMode.DRIVING to Transport.DRIVE,
-            TransportationMode.BICYCLING to Transport.BICYCLE,
-            TransportationMode.TRANSIT to Transport.TRANSIT
-        )
-    }
-
-    /**
-     * `Map<location_id, place>`
-     */
-    private val locationCache = mutableMapOf<String, Location>()
-
-    /**
-     * TODO ID가 좌표이면 즉시 인스턴스 생성해서 반환하기.
-     * TODO ID가 플레이스 ID이면 API에서 읽어서 캐시하고 반환하기.
-     */
-    override suspend fun read(id: String): Location? = locationCache[id]
-
-    override suspend fun searchDestination(center: GeoLocation, query: String): PlaceSearchResult {
-        logArgs(TAG, "searchDestination", "query" to query)
-
-        val now = timeProvider.instant()
-        val params = NearbySearchParams(
-            longitude = center.longitude,
-            latitude = center.latitude,
-            keyword = query,
-            rankBy = RankBy.DISTANCE
-        )
-        val apiResult = mapsClient.nearBy(params, now)
-
-        val result = PlaceSearchResult(
-            center, query,
-            apiResult.places.map { p ->
-                val location = LocationGmpPlace(p)
-                locationCache[location.id] = location
-                RecommendItemLocation(location)
-            },
-            apiResult.nextToken
-        )
-
-        Log.v(TAG, "#searchDestination return : $result")
-        return result
-    }
-
-    override suspend fun searchParking(destination: Location): PlaceSearchResult {
-        logArgs(TAG, "searchParking", "destination" to destination)
-
-        val now = timeProvider.instant()
-        val params = NearbySearchParams(
-            longitude = destination.longitude,
-            latitude = destination.latitude,
-            radius = SEARCH_PARKING_RADIUS,
-            type = PlaceTypes.PARKING
-        )
-        val apiResult = mapsClient.nearBy(params, now)
-        Log.v(TAG, "#searchParking : apiResult=$apiResult")
-
-        val result = PlaceSearchResult(
-            GeoLocation(destination.latitude, destination.longitude),
-            null,
-            apiResult.places.map { RecommendItemLocation(LocationGmpPlace(it)) },
-            apiResult.nextToken
-        )
-
-        Log.v(TAG, "#searchParking return : $result")
-        return result
-    }
-
-    private fun Location.toPlaceDescriptor() = when (this) {
-        is LocationGmpPlace -> PlaceDescriptor(placeId = place.placeId)
-        else -> PlaceDescriptor(geoLocation = GeoLocation(latitude, longitude))
-    }
-
-    override suspend fun searchRoute(
-        origin: Location,
-        destination: Location,
-        mode: TransportationMode
-    ): RouteSearchResult {
-        logArgs(TAG, "searchRoute", "origin" to origin, "destination" to destination, "mode" to mode)
-
-        val now = Instant.now()
-        val params = DirectionsParams(
-            origin = origin.toPlaceDescriptor(),
-            destination = destination.toPlaceDescriptor(),
-            transportationMode = mode
-        )
-        val result = mapsClient.directions(params, now)
-
-        val route = RouteSearchResult(
-            origin = origin,
-            destination = destination,
-            transport = TRANSPORTATION_MODE_MAP[mode]!!,
-            overview = result.routes[0].overviewPolyline.run {
-                Overview(points.map { GeoLocation(it.latitude, it.longitude) })
-            }
-        )
-        Log.v(TAG, "#searchRoute return : $route")
-        return route
-    }
-
-    override fun toString() = "$TAG(placesClient=$mapsClient, timeProvider=$timeProvider)"
-}
